+++ src/main/java/com/hotel/flint/user/employee/service/EmployeeRoomService.java
@@ -67,7 +67,7 @@ public class EmployeeRoomService {
     }
 
 //
-    public void modRoomPrice(Long id, Double newPrice) {
+    public void modRoomPrice(Long id, long newPrice) {
         Employee authenticatedEmployee = getAuthenticatedEmployee();
         if(!authenticatedEmployee.getDepartment().toString().equals("Room")){
             throw new IllegalArgumentException("접근 권한이 없습니다.");
